### 2017-08-16 #369 @arfon
👋 @sbinet. Many thanks for your submission to JOSS.

Please review the [author guidelines](http://joss.theoj.org/about#author_guidelines) and prepare a `paper.md` and `paper.bib` file and add them to your repository. Once you've got those in place we can start looking for an editor and reviewer for your submission.
