### 2017-08-16 #369 @sbinet
@arfon hello,

I put them in a branch+PR so I could incorporate all reviewers' comments in one go: go-hep/hep#88.

also, it wasn't clear to me whether they had to be at the root of the repo or inside their own directory (and, in that case, whether that directory's name was constrained by some convention).
so I put the `paper.*` files under a directory `joss-paper`, here:

 - https://github.com/go-hep/hep/blob/joss-paper/joss-paper/paper.md
 - https://github.com/go-hep/hep/blob/joss-paper/joss-paper/paper.bib
