### 2017-08-16 #369 @jpivarski
In fact, I'd like to review the go-hep code, as I've been meaning to go through it carefully at some point anyway. (I want to learn how you walk through ROOT headers, compared to RIO and root4j.) I have zero experience with Go, so I can't speak to a proper use of Go style, but I have been intending to learn it and this seems to be as good a reason as any.

Other than lack of familiarity with the language, I would be a good choice of reviewer because I'm familiar with the domain, the details of the ROOT file format, and I'm neutral: I don't have any competing projects or conflict of interest. (root4j is for a completely different purpose.) I didn't write any parts of go-hep, naturally.

It would have to be after ACAT (next week), though. There's just a few more conferences this year, so if this review is in the fall, that would be perfect.
