### 2017-08-17 21e8c23
rootio: reduce memory usage of branch/basket reading
benchmark                      old ns/op     new ns/op     delta
BenchmarkScanStruct-4          39159         38872         -0.73%
BenchmarkScanVars-4            37434         38277         +2.25%
BenchmarkScanVarsBigFile-4     110833139     118749880     +7.14%

benchmark                      old allocs     new allocs     delta
BenchmarkScanStruct-4          100            100            +0.00%
BenchmarkScanVars-4            100            100            +0.00%
BenchmarkScanVarsBigFile-4     727530         727502         -0.00%

benchmark  
[…]
