### 2017-08-17 660c24a
rootio: refactor test data type
