### 2017-08-17 a6bac13
rootio: add benchmarks for Scanner
BenchmarkScanStruct-4        	   50000	     39159 ns/op	     801 B/op	     100 allocs/op
BenchmarkScanVars-4          	   50000	     37434 ns/op	     800 B/op	     100 allocs/op
BenchmarkScanVarsBigFile-4   	      10	 110833139 ns/op	 8244192 B/op	  727530 allocs/op
PASS
ok  	go-hep.org/x/hep/rootio	5.910s
