### 2017-08-17 9e4946a
rootio: improve findBasketEntryIndex for big files
