### 2017-08-17 44499c9
rootio: simplify RBuffer internals
Replace RBuffer.rbuff with *bytes.Reader.

name               old time/op    new time/op    delta
ScanStruct-4         39.2µs ± 0%    35.8µs ± 1%    ~            (p=0.182 n=1+10)
ScanVars-4           37.4µs ± 0%    34.4µs ± 3%    ~            (p=0.182 n=1+10)
ScanVarsBigFile-4     111ms ± 0%     107ms ± 4%    ~            (p=0.364 n=1+10)

name               old alloc/op   new alloc/op   delta
ScanStruct-4           801B ± 0%      800B ± 0%    ~             (p=1.800 n=1+9)
ScanVars-4             800B ± 0%      800B ± 0%    ~     (all samples are equal)
ScanVa
[…]
