### 2017-08-17 bbc2531
rootio: use more efficient bytes reader
name               old time/op    new time/op    delta
ScanStruct-4         35.8µs ± 1%    29.7µs ± 0%   -17.15%   (p=0.000 n=10+8)
ScanVars-4           34.4µs ± 3%    27.4µs ± 0%   -20.32%  (p=0.000 n=10+10)
ScanVarsBigFile-4     107ms ± 4%      92ms ± 5%   -14.04%  (p=0.000 n=10+10)

name               old alloc/op   new alloc/op   delta
ScanStruct-4           800B ± 0%       0B ±NaN%  -100.00%   (p=0.000 n=9+10)
ScanVars-4             800B ± 0%       0B ±NaN%  -100.00%  (p=0.000 n=10+10)
ScanVarsBigFile-4    7.68MB ± 0%    5.87MB ± 0%   -23.58%   (p=0
[…]
