### 2017-08-17 b2e8205
rootio: removing stray line for godoc
