+++ rootio/cmd/root-srv/main_go16.go
@@ -19,7 +19,6 @@
 //
 //  $> root-srv -addr :8080 -serv https -host example.com
 //  2017/04/06 15:13:59 https server listening on :8080 at example.com
-
 package main
 
 import (
