### 2017-08-17 48dabe7
Adding mughetto ORCID to paper-md
