### 2017-08-17 a31fd0e
rootio: inline read-xyz operations of RBuffer
name               old time/op    new time/op    delta
ScanStruct-4         29.7µs ± 0%    27.7µs ± 0%   -6.59%         (p=0.000 n=8+10)
ScanVars-4           27.4µs ± 0%    25.7µs ± 0%   -6.40%         (p=0.000 n=10+9)
ScanVarsBigFile-4    92.0ms ± 5%    81.7ms ± 8%  -11.26%        (p=0.000 n=10+10)

name               old alloc/op   new alloc/op   delta
ScanStruct-4         0.00B ±NaN%    0.00B ±NaN%     ~     (all samples are equal)
ScanVars-4           0.00B ±NaN%    0.00B ±NaN%     ~     (all samples are equal)
ScanVarsBigFile-4    5.87MB ± 0% 
[…]
