### 2017-08-17 c5ef088
rootio: document rbuff, make it io.Reader/io.Seeker compliant
