### 2017-08-17 40f8371
rootio: rename Scanner into TreeReader, introduce fast Scanner
This CL renames Scanner into TreeReader and introduces a fast Scanner
that is early bound to a set of pointer to variables to fill.

name               old time/op    new time/op    delta
ScanStruct-4         27.9µs ± 0%    10.1µs ± 0%  -63.84%          (p=0.000 n=9+9)
ScanVars-4           25.2µs ± 0%    10.1µs ± 0%  -60.01%         (p=0.000 n=9+10)
ScanVarsBigFile-4    82.9ms ± 3%    38.2ms ± 8%  -53.90%        (p=0.000 n=10+10)

name               old alloc/op   new alloc/op   delta
ScanStruct-4         0.00B ±NaN%    0.00B ±NaN%
[…]
