+++ rootio/scanner.go
@@ -9,8 +9,7 @@ import (
 	"reflect"
 )
 
-// Scanner scans, selects and iterates over Tree entries.
-type Scanner struct {
+type baseScanner struct {
 	tree Tree
 	i    int64 // number of entries iterated over
 	n    int64 // number of entries to iterate over
@@ -21,10 +20,53 @@ type Scanner struct {
 	cbr []Branch    // branches activated because holding slice index
 	ibr []scanField // indices of activated branches
 
-	typ reflect.Type  // type bound to this scanner (a struct, a map, a slice of types)
-	ptr reflect.Value // pointer to value bound to this scanner
-
 	closed bool
+}
+
+// Close closes the Scanner, preventing further iteration.
+// Close is idempotent and does not affect the result of Err.
+func (s *baseScanner) Close() error {
+	if s.closed {
+		return nil
+	}
+	s.closed = true
+	s.tree = nil
+	s.mbr = nil
+	s.ibr = nil
+	return nil
+}
+
+// Err returns the error, if any, that was encountered during iteration.
+func (s *baseScanner) Err() error {
+	return s.err
+}
+
+// Entry returns the entry number of the last read row.
+func (s *baseScanner) Entry() int64 {
+	return s.cur
+}
+
+// SeekEntry points the scanner to the i-th entry, ready to call Next.
+func (s *baseScanner) SeekEntry(i int64) error {
+	if s.err != nil {
+		return s.err
+	}
+	s.i = i
+	s.cur = i - 1
+	return s.err
+}
+
+// Next prepares the next result row for reading with the Scan method.
+// It returns true on success, false if there is no next result row.
+// Every call to Scan, even the first one, must be preceded by a call to Next.
+func (s *baseScanner) Next() bool {
+	if s.closed {
+		return false
+	}
+	next := s.i < s.n
+	s.cur++
+	s.i++
+	return next
 }
 
 // scanField associates a Branch with a struct's field index
@@ -33,14 +75,23 @@ type scanField struct {
 	i  int // field index
 }
 
-// NewScanner creates a new Scanner connecting the pointer to some
+// TreeScanner scans, selects and iterates over Tree entries.
+type TreeScanner struct {
+	scan baseScanner
+
+	typ reflect.Type  // type bound to this scanner (a struct, a map, a slice of types)
+	ptr reflect.Value // pointer to value bound to this scanner
+}
+
+// NewTreeScanner creates a new Scanner connecting the pointer to some
 // user provided type to the given Tree.
-func NewScanner(t Tree, ptr interface{}) (*Scanner, error) {
+func NewTreeScanner(t Tree, ptr interface{}) (*TreeScanner, error) {
 	mbr := make([]Branch, 0, len(t.Branches()))
 	ibr := make([]scanField, 0, cap(mbr))
+	cbr := make([]Branch, 0)
 	rt := reflect.TypeOf(ptr).Elem()
 	if rt.Kind() != reflect.Struct {
-		return nil, errorf("rootio: NewScanner expects a pointer to a struct (got: %T)", ptr)
+		return nil, errorf("rootio: NewTreeScanner expects a pointer to a struct (got: %T)", ptr)
 	}
 	rv := reflect.New(rt).Elem()
 	for i := 0; i < rt.NumField(); i++ {
@@ -53,6 +104,14 @@ func NewScanner(t Tree, ptr interface{}) (*Scanner, error) {
 		if br == nil {
[…]
