+++ rootio/cmd/root-srv/server/plots.go
@@ -235,7 +235,7 @@ func (srv *server) plotBranchHandle(w http.ResponseWriter, r *http.Request) erro
 	min := +math.MaxFloat64
 	max := -math.MaxFloat64
 	vals := make([]float64, 0, int(tree.Entries()))
-	sc, err := rootio.NewScannerVars(tree, rootio.ScanVar{Name: leaf.Name()})
+	sc, err := rootio.NewTreeScannerVars(tree, rootio.ScanVar{Name: leaf.Name()})
 	if err != nil {
 		return fmt.Errorf("error creating scanner for branch %q in tree %q of file %q: %v", bname, tree.Name(), fname, err)
 	}
