+++ rootio/scanner_test.go
@@ -34,6 +34,83 @@ type ScannerData struct {
 	SliF64 []float64   `rootio:"SliceFloat64"`
 }
 
+func TestTreeScannerStruct(t *testing.T) {
+	f, err := Open("testdata/small-flat-tree.root")
+	if err != nil {
+		t.Fatal(err.Error())
+	}
+	defer f.Close()
+
+	obj, ok := f.Get("tree")
+	if !ok {
+		t.Fatalf("could not retrieve tree [tree]")
+	}
+	tree := obj.(Tree)
+
+	want := func(i int64) (data ScannerData) {
+		data.I32 = int32(i)
+		data.I64 = int64(i)
+		data.U32 = uint32(i)
+		data.U64 = uint64(i)
+		data.F32 = float32(i)
+		data.F64 = float64(i)
+		data.Str = fmt.Sprintf("evt-%03d", i)
+		for ii := range data.ArrI32 {
+			data.ArrI32[ii] = int32(i)
+			data.ArrI64[ii] = int64(i)
+			data.ArrU32[ii] = uint32(i)
+			data.ArrU64[ii] = uint64(i)
+			data.ArrF32[ii] = float32(i)
+			data.ArrF64[ii] = float64(i)
+		}
+		data.N = int32(i) % 10
+		data.SliI32 = make([]int32, int(data.N))
+		data.SliI64 = make([]int64, int(data.N))
+		data.SliU32 = make([]uint32, int(data.N))
+		data.SliU64 = make([]uint64, int(data.N))
+		data.SliF32 = make([]float32, int(data.N))
+		data.SliF64 = make([]float64, int(data.N))
+		for ii := 0; ii < int(data.N); ii++ {
+			data.SliI32[ii] = int32(i)
+			data.SliI64[ii] = int64(i)
+			data.SliU32[ii] = uint32(i)
+			data.SliU64[ii] = uint64(i)
+			data.SliF32[ii] = float32(i)
+			data.SliF64[ii] = float64(i)
+		}
+		return data
+	}
+
+	sc, err := NewTreeScanner(tree, &ScannerData{})
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer sc.Close()
+	var d1 ScannerData
+	for sc.Next() {
+		err := sc.Scan(&d1)
+		if err != nil {
+			t.Fatal(err)
+		}
+		i := sc.Entry()
+		if !reflect.DeepEqual(d1, want(i)) {
+			t.Fatalf("entry[%d]:\ngot= %#v.\nwant=%#v\n", i, d1, want(i))
+		}
+
+		var d2 ScannerData
+		err = sc.Scan(&d2)
+		if err != nil {
+			t.Fatal(err)
+		}
+		if !reflect.DeepEqual(d2, want(i)) {
+			t.Fatalf("entry[%d]:\ngot= %#v.\nwant=%#v\n", i, d2, want(i))
+		}
+	}
+	if err := sc.Err(); err != nil && err != io.EOF {
+		t.Fatal(err)
+	}
+}
+
 func TestScannerStruct(t *testing.T) {
 	f, err := Open("testdata/small-flat-tree.root")
 	if err != nil {
@@ -81,29 +158,30 @@ func TestScannerStruct(t *testing.T) {
 		return data
 	}
 
-	sc, err := NewScanner(tree, &ScannerData{})
+	var data ScannerData
+	sc, err := NewScanner(tree, &data)
 	if err != nil {
 		t.Fatal(err)
 	}
 	defer sc.Close()
-	var d1 ScannerData
 	for sc.Next() {
-		err := sc.Scan(&d1)
+		err := sc.Scan()
[…]
