+++ cmd/root2npy/main.go
@@ -188,7 +188,7 @@ func main() {
 	}
 	log.Printf("scanning leaves... [done]")
 
-	sc, err := rootio.NewScannerVars(tree, nt.args...)
+	sc, err := rootio.NewTreeScannerVars(tree, nt.args...)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -256,7 +256,7 @@ func (nt *ntuple) add(name string, leaf rootio.Leaf) {
 	n := len(nt.cols)
 	nt.cols = append(nt.cols, newColumn(name, leaf, nt.n))
 	col := &nt.cols[n]
-	nt.args = append(nt.args, rootio.ScanVar{Name: name, Type: col.etype})
+	nt.args = append(nt.args, rootio.ScanVar{Name: name})
 	nt.vars = append(nt.vars, col.data.Addr().Interface())
 }
 

[…]
