### 2017-08-17 f2074e6
rootio: add benchmark for big files w/ slice or scalar
