### 2017-08-18 5d6cafa
fastjet/internal/delaunay: implement hierarchical removal
BenchmarkHierarchicalDelaunayInsertionAndRemoval50-4                 100
  14990644 ns/op          251160 B/op       9868 allocs/op
BenchmarkHierarchicalDelaunayInsertionAndRemovall00-4                 50
  31802214 ns/op          519887 B/op      20575 allocs/op
BenchmarkHierarchicalDelaunayInsertionAndRemoval150-4                 30
  48335503 ns/op          789009 B/op      31358 allocs/op
BenchmarkHierarchicalDelaunayInsertionAndRemoval200-4                 20
  64054045 ns/op         1036826 B/op      41151 allocs/op
BenchmarkHiera
[…]
