+++ fastjet/internal/delaunay/delaunay_test.go
@@ -297,3 +297,278 @@ func BenchmarkHierarchicalDelaunayInsertion950(b *testing.B) {
 func BenchmarkHierarchicalDelaunayInsertion1000(b *testing.B) {
 	benchmarkHierarchicalDelaunayInsertion(1000, b)
 }
+
+func TestHierarchicalDelaunayRemovalSmall(t *testing.T) {
+	// NewPoint(x, y)
+	p1 := NewPoint(0, 0)
+	p2 := NewPoint(0, 2)
+	p3 := NewPoint(1, 0)
+	p4 := NewPoint(4, 4)
+	// point to be removed later
+	pE := NewPoint(3, 2)
+	ps := []*Point{
+		p1,
+		p2,
+		p3,
+		pE,
+		p4,
+	}
+	d := HierarchicalDelaunay()
+	for _, p := range ps {
+		d.Insert(p)
+	}
+	d.Remove(pE)
+	exp := []*Triangle{
+		NewTriangle(p1, p2, p3),
+		NewTriangle(p2, p3, p4),
+	}
+	ts := d.Triangles()
+	got, want := len(ts), len(exp)
+	if got != want {
+		t.Errorf("got=%d delaunay triangles, want=%d", got, want)
+	}
+	for i := range ts {
+		ok := false
+		for j := range exp {
+			if ts[i].Equals(exp[j]) {
+				ok = true
+				// remove triangles that have been matched from slice,
+				// in case there are duplicate triangles.
+				exp = append(exp[:j], exp[j+1:]...)
+				break
+			}
+		}
+		if !ok {
+			t.Errorf("Triangle T%s not as expected", ts[i])
+		}
+	}
+	var (
+		nn []*Point
+		nd []float64
+	)
+	for _, p := range ps {
+		if p.Equals(pE) {
+			continue
+		}
+		n, d := p.NearestNeighbor()
+		nn = append(nn, n)
+		nd = append(nd, d)
+	}
+	expN := []*Point{p3, p1, p1, p2}
+	expD := []float64{1.0, 2.0, 1.0, 4.4721}
+	got, want = len(nn), len(expN)
+	if got != want {
+		t.Errorf("got=%d nearest neighbors, want=%d", got, want)
+	}
+	for i := range nn {
+		if !nn[i].Equals(expN[i]) {
+			t.Errorf("got=N%s nearest neighbor, want=N%s", nn[i], expN[i])
+		}
+		if math.Abs(nd[i]-expD[i]) > tol {
+			t.Errorf("got=%f distance, want=%f for point P%s with neighbour N%s", nd[i], expD[i], ps[i], nn[i])
+		}
+	}
+}
+
+func TestHierarchicalDelaunayRemovalMedium(t *testing.T) {
+	// NewPoint(x, y)
+	p1 := NewPoint(-1.5, 3.2)
+	p2 := NewPoint(1.8, 3.3)
+	p3 := NewPoint(-3.7, 1.5)
+	p4 := NewPoint(-1.5, 1.3)
+	p5 := NewPoint(0.8, 1.2)
+	p6 := NewPoint(3.3, 1.5)
+	p7 := NewPoint(-4, -1)
+	p8 := NewPoint(-2.3, -0.7)
+	p9 := NewPoint(0, -0.5)
+	p10 := NewPoint(2, -1.5)
+	p11 := NewPoint(3.7, -0.8)
+	p12 := NewPoint(-3.5, -2.9)
+	p13 := NewPoint(-0.9, -3.9)
+	p14 := NewPoint(2, -3.5)
+	p15 := NewPoint(3.5, -2.25)
+	// points to be removed later
+	pE1 := NewPoint(0, 0)
+	pE2 := NewPoint(-2.3, -0.6)
[…]
