### 2017-08-19 #369 @arfon
> I put them in a branch+PR so I could incorporate all reviewers' comments in one go: go-hep/hep#88.

Got it. Thanks.
