### 2017-08-19 #372 @whedon
[REVIEW]: Go-HEP: libraries for High Energy Physics analyses in Go
**Submitting author:** @sbinet (<a href="http://orcid.org/0000-0003-4913-6104">Sebastien Binet</a>)
**Repository:** <a href="https://github.com/go-hep/hep" target ="_blank">https://github.com/go-hep/hep</a>
**Version:** v0.4
