### 2017-08-19 #372 @jpivarski
What are the deadlines for this review (rough or strict)? I need to schedule it in with other activities.
