### 2017-08-19 #372 @arfon
> What are the deadlines for this review (rough or strict)? I need to schedule it in with other activities.

Ideally in the next couple of weeks. We don't have strict deadlines.
