### 2017-08-22 fe5f91e
hbook: rename annYODA to annToYODA
