### 2017-08-22 bd7b2f3
hbook: introduce annFromYODA
