### 2017-08-22 72eff52
hbook: properly handle 'name' annotation wrt YODA
