+++ hbook/h2d.go
@@ -219,6 +219,9 @@ func (h *H2D) annToYODA() Annotation {
 	ann["Path"] = "/" + h.Name()
 	ann["Title"] = ""
 	for k, v := range h.ann {
+		if k == "name" {
+			continue
+		}
 		ann[k] = v
 	}
 	return ann
