+++ fastjet/internal/heap/doc.go
@@ -1,0 +1,6 @@
+// Copyright 2017 The go-hep Authors.  All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+// package heap implements a min-heap for pairs of jets.
+package heap
