### 2017-08-23 47d64b0
fastjet/internal/heap: new package to implement a min heap for pairs of jets
BenchmarkHeap-4             1000           2082086 ns/op          350832 B/op
      17 allocs/op
BenchmarkPQ-4                300           4735609 ns/op           78984 B/op
      17 allocs/op
