### 2017-08-24 d3a7b82
fastjet:  protect against parton-level events where pt can be zero for some partons, giving rapidity=infinity
Fixes go-hep/hep#97
