### 2017-08-26 81ebe45
fastjet/internal/delaunay: remove root points from updatedNearestNeighbor & assign ID incrementally
