+++ fastjet/internal/delaunay/delaunay.go
@@ -20,6 +20,7 @@ type Delaunay struct {
 	// to locate a point. The variable root's nil-ness also indicates which method to use to locate a point.
 	root *Triangle
 	r    *rand.Rand
+	n    int // n is the number of points inserted. It is used to assign the id to points.
 }
 
 // HierarchicalDelaunay creates a Delaunay Triangulation using the delaunay hierarchy.
@@ -42,6 +43,7 @@ func HierarchicalDelaunay() *Delaunay {
 	root := NewTriangle(a, b, c)
 	return &Delaunay{
 		root: root,
+		n:    0,
 	}
 }
 
@@ -69,18 +71,23 @@ func (d *Delaunay) Triangles() []*Triangle {
 // whose nearest neighbor changed due to the insertion. The slice may contain
 // duplicates.
 func (d *Delaunay) Insert(p *Point) (updatedNearestNeighbor []*Point) {
+	p.id = d.n
+	d.n++
 	if len(p.adjacentTriangles) == 0 {
 		p.adjacentTriangles = make(triangles, 0)
 	}
 	p.adjacentTriangles = p.adjacentTriangles[:0]
 	t, l := d.locatePointHierarchy(p, d.root)
+	var updated points
 	switch l {
 	case inside:
-		return d.insertInside(p, t)
+		updated = d.insertInside(p, t)
 	case onEdge:
-		return d.insertOnEdge(p, t)
+		updated = d.insertOnEdge(p, t)
+	default:
+		panic(fmt.Errorf("delaunay: no triangle containing point %v", p))
 	}
-	panic(fmt.Errorf("delaunay: no triangle containing point %v", p))
+	return updated.remove(d.root.A, d.root.B, d.root.C)
 }
 
 // Remove removes the point from the triangulation. It returns the points
@@ -90,15 +97,16 @@ func (d *Delaunay) Remove(p *Point) (updatedNearestNeighbor []*Point) {
 	if len(p.adjacentTriangles) < 3 {
 		panic(fmt.Errorf("delaunay: can't remove point %v, not enough adjacent triangles", p))
 	}
-	points := p.surroundingPoints()
+	var updated points
+	pts := p.surroundingPoints()
 	ts := make([]*Triangle, len(p.adjacentTriangles))
 	copy(ts, p.adjacentTriangles)
 	for _, t := range ts {
 		updtemp := t.remove()
-		updatedNearestNeighbor = append(updatedNearestNeighbor, updtemp...)
+		updated = append(updated, updtemp...)
 	}
-	updtemp := d.retriangulateAndSew(points, ts)
-	return append(updatedNearestNeighbor, updtemp...)
+	updtemp := d.retriangulateAndSew(pts, ts)
+	return append(updated, updtemp...).remove(d.root.A, d.root.B, d.root.C)
 }
 
 // locatePointHierarchy locates the point using the delaunay hierarchy.
@@ -392,7 +400,6 @@ func (d *Delaunay) retriangulateAndSew(points []*Point, parents []*Triangle) (up
 	copies := make([]*Point, len(points))
 	for i, p := range points {
 		copies[i] = NewPoint(p.x, p.y)
-		copies[i].id = i
 		nd.Insert(copies[i])
 	}
 	ts := nd.Triangles()
