### 2017-08-26 633f5d2
fastjet/internal/delaunay: remove root points from updatedNearestNeighbor & assign ID incrementally
