### 2017-08-27 c8889e8
cmd/{rio,root}2yoda, cmd/yoda2rio: handle gzipped YODA files
This CL adds support for gzipped YODA files.
If the input or output YODA file name ends with ".gz", the appropriate
gzip.Reader or gzip.Writer is inserted in the stack of readers or
writers.
