### 2017-08-27 daaaf45
cmd/{rio,root}2yoda, cmd/yoda2rio: handle gzipped YODA files
This CL adds support for gzipped YODA files.
If the input or output YODA file name ends with ".gz", the appropriate
gzip.Reader or gzip.Writer is inserted in the stack of readers or
writers.
