### 2017-08-27 33f4273
hep: drop 1.6.x from travis, add 1.9.x
