+++ .travis.yml
@@ -1,9 +1,9 @@
 language: go
 go_import_path: go-hep.org/x/hep
 go:
-  - 1.6.x
   - 1.7.x
   - 1.8.x
+  - 1.9.x
   - master
 
 matrix:
