### 2017-08-28 e20807c
fastjet: integrate delaunay into clustersequence
