+++ fastjet/clustersequence.go
@@ -9,6 +9,7 @@ import (
 	"fmt"
 	"math"
 
+	"go-hep.org/x/hep/fastjet/internal/delaunay"
 	"go-hep.org/x/hep/fastjet/internal/heap"
 	"go-hep.org/x/hep/fmom"
 )
@@ -229,7 +230,7 @@ func (cs *ClusterSequence) run() error {
 	}
 
 	// FIXME
-	err = cs.runN3Dumb()
+	err = cs.runNlnN()
 	if err != nil {
 		return err
 	}
@@ -506,7 +507,164 @@ func (cs *ClusterSequence) runN3Dumb() error {
 // There are internally asserted assumptions about absence of points
 // with coincident eta-phi coordinates.
 func (cs *ClusterSequence) runNlnN() error {
-	panic(fmt.Errorf("fastjet: runNlnN not implemented"))
+	d := delaunay.HierarchicalDelaunay()
+	points := make([]pointInfo, len(cs.jets), 2*len(cs.jets))
+	// create points
+	for i := 0; i < len(cs.jets); i++ {
+		rap := cs.jets[i].Rapidity()
+		phi := cs.jets[i].Phi()
+		for phi < 0 {
+			phi += math.Pi * 2
+		}
+		for phi >= math.Pi*2 {
+			phi -= math.Pi * 2
+		}
+		p := delaunay.NewPoint(rap, phi)
+		d.Insert(p)
+		points[i] = pointInfo{p: p, jet: i, mirror: -1}
+	}
+	for i := range points {
+		_, dij := points[i].p.NearestNeighbor()
+		x, y := points[i].p.Coordinates()
+		// when a nearest neighbor is farther than the border a mirror point is inserted.
+		if dij > y {
+			y += 2 * math.Pi
+			p := delaunay.NewPoint(x, y)
+			points = append(points, pointInfo{p: p, jet: points[i].jet, mirror: i, clustered: false})
+			d.Insert(p)
+			points[i].mirror = len(points) - 1
+		} else if dij > 2*math.Pi-y {
+			y -= 2 * math.Pi
+			p := delaunay.NewPoint(x, y)
+			points = append(points, pointInfo{p: p, jet: points[i].jet, mirror: i, clustered: false})
+			d.Insert(p)
+			points[i].mirror = len(points) - 1
+		}
+	}
+	h := heap.New()
+	// insert the nearest neighbors of the initial points into the heap.
+	for i := range points {
+		n, dij := points[i].p.NearestNeighbor()
+		j := n.ID()
+		cs.addKtDistance(h, points, i, j, dij)
+		if mirror := points[i].mirror; mirror != -1 {
+			n, dij = points[mirror].p.NearestNeighbor()
+			j = n.ID()
+			cs.addKtDistance(h, points, mirror, j, dij)
+		}
+	}
+	n := len(cs.jets)
+	for i := 0; i < n; i++ {
+		var jeti, jetj int
+		var dij float64
+		var recombineWithBeam bool
+		clusteredi, clusteredj := true, true
+		// find a pair that hasn't recombined yet.
+		for clusteredi || clusteredj {
+			jeti, jetj, dij = h.Pop()
+			if jeti == -1 {
+				panic(fmt.Errorf("fastjet: heap is empty at index %d", i))
+			}
+			clusteredi = points[jeti].clustered
+			recombineWithBeam = (jetj == beamJetIndex)
+			if recombineWithBeam {
+				clusteredj = false
+			} else {
+				clusteredj = points[jetj].clustered
+			}
+		}
+		var updated []*delaunay.Point
+		if recombineWithBeam {
+			points[jeti].clustered = true
[…]
