### 2017-08-29 d9a852f
fastjet: add dispatch code for Best|N3Dumb|NlnN strategies
