### 2017-08-29 4d94aff
fastjet: add strategies test for simple data
