+++ fastjet/fastjet_test.go
@@ -5,6 +5,7 @@
 package fastjet_test
 
 import (
+	"fmt"
 	"sort"
 	"testing"
 
@@ -23,26 +24,20 @@ func TestSimple(t *testing.T) {
 		fastjet.NewJet(+50.0, +1.0, 100, 399.0),
 	}
 
-	// for i, jet := range particles {
-	// 	fmt.Printf("part[%d]: pt=%+e eta=%+e rap=%+e phi=%+e\n",
-	// 		i, jet.Pt(), jet.Eta(), jet.Rapidity(), jet.Phi(),
-	// 	)
-	// }
-
 	// choose a jet definition
 	r := 0.7
-	def := fastjet.NewJetDefinition(fastjet.AntiKtAlgorithm, r, fastjet.EScheme, fastjet.BestStrategy)
+	def := fastjet.NewJetDefinition(fastjet.AntiKtAlgorithm, r, fastjet.EScheme, fastjet.N3DumbStrategy)
 
 	if def.R() != r {
 		t.Fatalf("expected r-param=%v. got=%v", r, def.R())
 	}
 
 	if def.RecombinationScheme() != fastjet.EScheme {
-		t.Fatalf("expected scheme=%v. got=%v", def.RecombinationScheme(), fastjet.EScheme)
+		t.Fatalf("got scheme=%v. want=%v", def.RecombinationScheme(), fastjet.EScheme)
 	}
 
-	if def.Strategy() != fastjet.BestStrategy {
-		t.Fatalf("expected strategy=%v. got=%v", def.Strategy(), fastjet.BestStrategy)
+	if def.Strategy() != fastjet.N3DumbStrategy {
+		t.Fatalf("got strategy=%v. want=%v", def.Strategy(), fastjet.N3DumbStrategy)
 	}
 
 	// run the clustering, extract jets
@@ -110,6 +105,159 @@ func TestSimple(t *testing.T) {
 					jj.PxPyPzE,
 				)
 			}
+		}
+	}
+}
+
+func TestStrategies(t *testing.T) {
+	type result struct {
+		jet fmom.PxPyPzE
+		cts []fmom.PxPyPzE
+	}
+
+	for _, test := range []struct {
+		name       string
+		particles  []fastjet.Jet
+		r          float64
+		scheme     fastjet.RecombinationScheme
+		ptmin      float64
+		strategies []fastjet.Strategy
+		want       []result
+	}{
+		{
+			name: "simple-1",
+			particles: []fastjet.Jet{
+				fastjet.NewJet(+99.0, +0.1, 0, 100.0),
+				fastjet.NewJet(+04.0, -0.1, 0, 005.0),
+				fastjet.NewJet(-99.0, +0.0, 0, 099.0),
+				fastjet.NewJet(+99.0, +0.1, 0, 199.0),
+				fastjet.NewJet(-99.0, +0.0, 0, 299.0),
+				fastjet.NewJet(-99.0, +1.0, 0, 399.0),
+				fastjet.NewJet(+50.0, +1.0, 100, 399.0),
+			},
+			r:          0.7,
+			scheme:     fastjet.EScheme,
+			ptmin:      0,
+			strategies: []fastjet.Strategy{fastjet.BestStrategy, fastjet.N3DumbStrategy /*FIXME(sbinet): add fastjet.NlnNStrategy*/},
+			want: []result{
+				{
+					jet: fmom.NewPxPyPzE(-2.970000e+02, +1.000000e+00, +0.000000e+00, +7.970000e+02),
+					cts: []fmom.PxPyPzE{
+
+						fmom.NewPxPyPzE(-9.900000e+01, +1.000000e+00, +0.000000e+00, +3.990000e+02),
[…]
