### 2017-08-29 c6c7d81
fastjet/internal/delaunay: handle duplicate points
