### 2017-08-29 76ba768
fastjet/internal/delaunay: handle duplicate points
