### 2017-08-30 742c651
fit: add numerical computation of hessian
