### 2017-08-30 821e4f4
all: use gonum.org/v1/plot import path
