+++ rootio/cmd/root-srv/server/data.go
@@ -11,9 +11,9 @@ import (
 	"strings"
 	"sync"
 
-	"github.com/gonum/plot/vg"
-	"github.com/gonum/plot/vg/draw"
-	"github.com/gonum/plot/vg/vgsvg"
+	"gonum.org/v1/plot/vg"
+	"gonum.org/v1/plot/vg/draw"
+	"gonum.org/v1/plot/vg/vgsvg"
 
 	"go-hep.org/x/hep/hplot"
 	"go-hep.org/x/hep/rootio"
