+++ lcio/example/lcio-ex-read-event/main.go
@@ -20,10 +20,10 @@ import (
 	"log"
 	"os"
 
-	"github.com/gonum/plot/vg"
 	"go-hep.org/x/hep/hbook"
 	"go-hep.org/x/hep/hplot"
 	"go-hep.org/x/hep/lcio"
+	"gonum.org/v1/plot/vg"
 )
 
 func main() {
