+++ hplot/hplot.go
@@ -13,12 +13,12 @@ import (
 
 	"golang.org/x/exp/shiny/screen"
 
-	"github.com/gonum/plot"
-	"github.com/gonum/plot/plotter"
-	"github.com/gonum/plot/vg"
-	"github.com/gonum/plot/vg/draw"
-	"github.com/gonum/plot/vg/vgimg"
 	"go-hep.org/x/hep/hplot/vgshiny"
+	"gonum.org/v1/plot"
+	"gonum.org/v1/plot/plotter"
+	"gonum.org/v1/plot/vg"
+	"gonum.org/v1/plot/vg/draw"
+	"gonum.org/v1/plot/vg/vgimg"
 )
 
 // Plot is the basic type representing a plot.
