### 2017-08-30 d67ae71
all: use gonum.org/v1/plot import path
