+++ fit/run.go
@@ -12,9 +12,9 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/gonum/plot"
-	"github.com/gonum/plot/plotter"
-	"github.com/gonum/plot/vg"
+	"gonum.org/v1/plot"
+	"gonum.org/v1/plot/plotter"
+	"gonum.org/v1/plot/vg"
 	"go-hep.org/x/hep/fit"
 	"go-hep.org/x/hep/hplot"
 	"gonum.org/v1/gonum/optimize"
