+++ fit/run.go
@@ -12,12 +12,12 @@ import (
 	"strconv"
 	"strings"
 
-	"gonum.org/v1/plot"
-	"gonum.org/v1/plot/plotter"
-	"gonum.org/v1/plot/vg"
 	"go-hep.org/x/hep/fit"
 	"go-hep.org/x/hep/hplot"
 	"gonum.org/v1/gonum/optimize"
+	"gonum.org/v1/plot"
+	"gonum.org/v1/plot/plotter"
+	"gonum.org/v1/plot/vg"
 )
 
 func fitFunc(x float64, p []float64) float64 {
