### 2017-08-31 6b9ebcf
fastjet/internal/delaunay: add function SecondNearestNeighbor
