+++ fastjet/internal/plot/doc.go
@@ -1,0 +1,6 @@
+// Copyright 2017 The go-hep Authors.  All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+// package plot contains functions to plot the Delaunay Triangulation and the Voronoi Diagram.
+package plot
