### 2017-09-07 d143233
all: add copyright headers
