+++ fwk/cmd/fwk-new-comp/tsk_template.go
@@ -1,3 +1,7 @@
+// Copyright 2017 The go-hep Authors.  All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
 package main
 
 const g_task_template = `package {{.Package}}
