+++ hepmc/go-hepmc-dump/main.go
@@ -1,3 +1,7 @@
+// Copyright 2017 The go-hep Authors.  All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
 // go-hepmc-dump is a simple command to dump in an almost human-friendly format
 // the content of a hepmc file.
 // ex:
