+++ hepmc/filetype.go
@@ -1,3 +1,7 @@
+// Copyright 2017 The go-hep Authors.  All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
 package hepmc
 
 const (
@@ -25,5 +29,3 @@ const (
 	hepmcASCIIPdt
 	hepmcExtendedASCIIPdt
 )
-
-// EOF
