+++ fwk/svcbase.go
@@ -1,3 +1,7 @@
+// Copyright 2017 The go-hep Authors.  All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
 package fwk
 
 import (
@@ -53,5 +57,3 @@ func (svc *SvcBase) GetProp(name string) (interface{}, error) {
 func (svc *SvcBase) FSMState() fsm.State {
 	return svc.mgr.FSMState()
 }
-
-// EOF
