+++ fwk/taskbase.go
@@ -1,3 +1,7 @@
+// Copyright 2017 The go-hep Authors.  All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
 package fwk
 
 import (
@@ -65,5 +69,3 @@ func (tsk *TaskBase) GetProp(n string) (interface{}, error) {
 func (tsk *TaskBase) FSMState() fsm.State {
 	return tsk.mgr.FSMState()
 }
-
-// EOF
