+++ hepmc/version.go
@@ -1,8 +1,10 @@
+// Copyright 2017 The go-hep Authors.  All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
 package hepmc
 
 // VersionName returns the hepmc version string.
 func VersionName() string {
 	return "2.06.09"
 }
-
-// EOF
