+++ fwk/io.go
@@ -1,3 +1,7 @@
+// Copyright 2017 The go-hep Authors.  All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
 package fwk
 
 // StreamControl provides concurrency-safe control to input and output streamers.
@@ -43,5 +47,3 @@ type OutputStreamer interface {
 	// It does not (and can not) close the underlying io.Writer.
 	Disconnect() error
 }
-
-// EOF
