### 2017-09-07 80b9335
fads/cmd/fads-app: add documentation
