### 2017-09-07 01f4f1d
cmd/{rio,root}2yoda, cmd/yoda2rio: add handling of .yodz compressed files
