### 2017-09-07 #372 @jpivarski
I don't think the `LICENSE` file is sufficient. It claims copyright and states terms for redistribution, but (not being a lawyer) I don't know if that covers the full terms of an [OSI license](https://opensource.org/licenses/alphabetical). Just pick a common one (ASF2, BSD, MIT).

Still going through the other steps, from top to bottom of the checklist...
