### 2017-09-07 #372 @jpivarski
Nevermind! I see on [https://go-hep.org/](go-hep.org) that it's licensed under BSD-3, so I was about to say, "copy-paste the text of a BSD-3" license, but that's exactly what he did. It just happens to be very short.

So the license is fine.
