### 2017-09-07 #372 @jpivarski
Installation _seemed_ to go without problems. I'm not getting error messages from the import statement, so I think I can check off that box.
