### 2017-09-07 #372 @jpivarski
There are clearly automated tests because Travis finds them, but I can't find them. The `.travis.yml` is so basic they must be in the standard place for Go, but I don't know what that is. Based on the fact that I see clear and detailed Travis reports, I'm going to check this box.
