### 2017-09-07 #372 @sbinet
yes, travis knows how Go tests are run.

if you want to run them by hand:

```sh
$> cd $GOPATH/src/go-hep.org/x/hep
$> go test -v ./...
```

(yes, with the `...` ellipsis at the end: this instructs the Go toolchain to recurse and traverse all the tree under the current working directory)
