### 2017-09-07 #372 @jpivarski
There's also clear community support guidelines and good API documentation.

However, I'm not able to find any tutorials for getting started. I picked x/hep/rootio as a starting point (my primary interest) and I can't get it to work. I'll submit an issue and see what happens.
