### 2017-09-07 #372 @sbinet
Go has a nice "examples" feature that is leveraged by the documentation presentation tool, godoc.

Examples for a type inside a package are _e.g._ here:

- https://godoc.org/go-hep.org/x/hep/rootio#pkg-examples
