### 2017-09-07 1386a05
rootio: update package doc, reference rootio examples
The package doc was still refering to the old way to retrieve values
from a rootio.Directory.
Fix that and expand it a bit.

Fixes go-hep/hep#112.
