### 2017-09-07 #372 @sbinet
and if you want to run benchmarks:

```
$> cd rootio
$> go test -v -run=NONE -bench=. -benchmem -count=5 .
goos: linux
goarch: amd64
pkg: go-hep.org/x/hep/rootio
BenchmarkTreeScannerStruct-4              	  100000	     17515 ns/op	       0 B/op	       0 allocs/op
BenchmarkTreeScannerStruct-4              	  100000	     17521 ns/op	       0 B/op	       0 allocs/op
BenchmarkTreeScannerStruct-4              	  100000	     17437 ns/op	       0 B/op	       0 allocs/op
BenchmarkTreeScannerStruct-4              	  100000	     17514 ns/op	       0 B/op	       0 allocs/op
BenchmarkTreeScannerStruct-4              	  100000	     17562 ns/op	       0 B/op	       0 allocs/op
BenchmarkScannerStruct-4                  	  200000	      6209 ns/op	       0 B/op	       0 allocs/op
BenchmarkScannerStruct-4                  	  200000	      6253 ns/op	       0 B/op	       0 allocs/op
BenchmarkScannerStruct-4                  	  200000	      6257 ns/op	       0 B/op	       0 allocs/op
BenchmarkScannerStruct-4                  	  200000	      6257 ns/op	       0 B/op	       0 allocs/op
BenchmarkScannerStruct-4                  	  200000	      6226 ns/op	       0 B/op	       0 allocs/op
BenchmarkTreeScannerVars-4                	  100000	     15960 ns/op	       0 B/op	       0 allocs/op
BenchmarkTreeScannerVars-4                	  100000	     15952 ns/op	       0 B/op	       0 allocs/op
BenchmarkTreeScannerVars-4                	  100000	     15961 ns/op	       0 B/op	       0 allocs/op
BenchmarkTreeScannerVars-4                	  100000	     15913 ns/op	       0 B/op	       0 allocs/op
BenchmarkTreeScannerVars-4                	  100000	     16182 ns/op	       0 B/op	       0 allocs/op
BenchmarkScannerVars-4                    	  200000	      6116 ns/op	       0 B/op	       0 allocs/op
BenchmarkScannerVars-4                    	  200000	      6138 ns/op	       0 B/op	       0 allocs/op
BenchmarkScannerVars-4                    	  200000	      6169 ns/op	       0 B/op	       0 allocs/op
BenchmarkScannerVars-4                    	  200000	      6117 ns/op	       0 B/op	       0 allocs/op
BenchmarkScannerVars-4                    	  200000	      6113 ns/op	       0 B/op	       0 allocs/op
BenchmarkTreeScannerVarsBigFileScalar-4   	     100	  22425657 ns/op	  505110 B/op	  125000 allocs/op
BenchmarkTreeScannerVarsBigFileScalar-4   	     100	  22302968 ns/op	  505110 B/op	  125000 allocs/op
BenchmarkTreeScannerVarsBigFileScalar-4   	     100	  22443355 ns/op	  505111 B/op	  125000 allocs/op
BenchmarkTreeScannerVarsBigFileScalar-4   	     100	  22367911 ns/op	  505110 B/op	  125000 allocs/op
BenchmarkTreeScannerVarsBigFileScalar-4   	     100	  22492822 ns/op	  505110 B/op	  125000 allocs/op
BenchmarkScannerVarsBigFileScalar-4       	     200	   7702788 ns/op	    2554 B/op	       0 allocs/op
BenchmarkScannerVarsBigFileScalar-4       	     200	   7749957 ns/op	    2554 B/op	       0 allocs/op
BenchmarkScannerVarsBigFileScalar-4       	     200	   7706357 ns/op	    2554 B/op	       0 allocs/op
BenchmarkScannerVarsBigFileScalar-4       	     200	   7651564 ns/op	    2554 B/op	       0 allocs/op
BenchmarkScannerVarsBigFileScalar-4       	     200	   7662549 ns/op	    2554 B/op	       0 allocs/op
BenchmarkTreeScannerVarsBigFileSlice-4    	      30	  41926597 ns/op	 5834710 B/op	  375033 allocs/op
BenchmarkTreeScannerVarsBigFileSlice-4    	      30	  40863289 ns/op	 5834708 B/op	  375033 allocs/op
BenchmarkTreeScannerVarsBigFileSlice-4    	      30	  41548413 ns/op	 5834724 B/op	  375034 allocs/op
BenchmarkTreeScannerVarsBigFileSlice-4    	      30	  41602640 ns/op	 5834708 B/op	  375033 allocs/op
BenchmarkTreeScannerVarsBigFileSlice-4    	      30	  41578124 ns/op	 5834720 B/op	  375033 allocs/op
BenchmarkScannerVarsBigFileSlice-4        	     100	  20647574 ns/op	 1287076 B/op	  157616 allocs/op
BenchmarkScannerVarsBigFileSlice-4        	     100	  20287316 ns/op	 1287075 B/op	  157616 allocs/op
BenchmarkScannerVarsBigFileSlice-4        	     100	  20401387 ns/op	 1287077 B/op	  157616 allocs/op
BenchmarkScannerVarsBigFileSlice-4        	     100	  20292135 ns/op	 1287076 B/op	  157616 allocs/op
BenchmarkScannerVarsBigFileSlice-4        	     100	  20671081 ns/op	 1287075 B/op	  157616 allocs/op
PASS
ok  	go-hep.org/x/hep/rootio	79.667s
```

`count=5` is to run each benchmark 5 times (to estimate std-err).

another tool [benchstat](https://godoc.org/rsc.io/benchstat) understands the output format of `go test -bench` and can extract meaningful statistics:

```
$> go get rsc.io/benchstat
$>  benchstat ./foo.txt
name                            time/op
TreeScannerStruct-4             17.5µs ± 0%
ScannerStruct-4                 6.24µs ± 1%
TreeScannerVars-4               15.9µs ± 0%
ScannerVars-4                   6.13µs ± 1%
TreeScannerVarsBigFileScalar-4  22.4ms ± 0%
ScannerVarsBigFileScalar-4      7.69ms ± 1%
TreeScannerVarsBigFileSlice-4   41.5ms ± 2%
ScannerVarsBigFileSlice-4       20.5ms ± 1%

name                            alloc/op
TreeScannerStruct-4              0.00B
ScannerStruct-4                  0.00B
TreeScannerVars-4                0.00B
ScannerVars-4                    0.00B
TreeScannerVarsBigFileScalar-4   505kB ± 0%
ScannerVarsBigFileScalar-4      2.55kB ± 0%
TreeScannerVarsBigFileSlice-4   5.83MB ± 0%
ScannerVarsBigFileSlice-4       1.29MB ± 0%

name                            allocs/op
TreeScannerStruct-4               0.00
ScannerStruct-4                   0.00
TreeScannerVars-4                 0.00
ScannerVars-4                     0.00
TreeScannerVarsBigFileScalar-4    125k ± 0%
ScannerVarsBigFileScalar-4        0.00
TreeScannerVarsBigFileSlice-4     375k ± 0%
ScannerVarsBigFileSlice-4         158k ± 0%
```

(and `benchstat` can compare 2 files and thus provide p-values between 2 sets of benchmarks)
