### 2017-09-07 #372 @jpivarski
I only needed a little orientation, so I'm going to say, "yes" to documentation. What I needed was there, I just couldn't find it. "Yes" to functionality because I could run it over one of my own data files (though I didn't do a comprehensive study, just a simple example).

And finally, "yes" to performance. Again, I haven't done a systematic study and my laptop has a weird configuration (no L3 cache in my CPU). However, I found go-hep/rootio to be only 25% slower than C++ ROOT on a normal, compressed file and _twice as fast as C++_ on an uncompressed file.

Hint, hint to the authors: include uncompressed files in your tests! Maybe you're only slow because the Go zlib is slow, and when you remove that, you expose a lower overhead in entry handling than `TTree::GetEntry`. ROOT is moving to LZ4 compression (by default), which is much faster than zlib, so this might not matter, as long as you can get LZ4 in Go.
