### 2017-09-07 #372 @jpivarski
Only a few checkboxes left: statement of need. The go-hep/hep README doesn't say why you might need such a thing. I think you can make an argument for it in one or two sentences.

Also, I don't remember where `paper.md` is. Could you point me to it again?

The ROOT team is in flux about what DOI they want to represent ROOT, now that there are options for referencing software, rather than just papers. Traditionally, the correct reference is their 1997 paper.
