### 2017-09-07 #372 @sbinet
`paper.md` is in that PR: go-hep/hep#88.
https://github.com/go-hep/hep/pull/88/files
