### 2017-09-07 #372 @jpivarski
The `paper.md` has a good statement of need. I'd say copy-paste a lot of that text into the README— it's a good introduction to the purpose of this project.

The ROOT reference you have is from 2009. It looks relevant, but I'm pretty sure they have a preference for the 1997 one, just so that they can count references.
