### 2017-09-07 #372 @jpivarski
It's a 1996 paper:

  - [https://root-forum.cern.ch/t/roots-references-in-a-paper/14638](https://root-forum.cern.ch/t/roots-references-in-a-paper/14638)
  - bottom of [https://root.cern.ch/root/Publications.html](https://root.cern.ch/root/Publications.html)
