### 2017-09-08 3a13c67
fastjet/internal/delaunay: implement voronoi
