### 2017-09-08 8d5e7a1
fastjet/internal/delaunay: implement voronoi
