### 2017-09-08 #372 @sbinet
PTAL.

- https://github.com/go-hep/hep/pull/88/commits/d438a0801e59aedb618a87a751d0bab111cf8e96 (ROOT paper)
- https://github.com/go-hep/hep/pull/88/commits/936377e5f7dd9586bc0e46d237893027ce20609f (add motivations to `README`)
