### 2017-09-08 #372 @sbinet
also: @jpivarski  actually reviewed (implicitly) `master` which is more or less what `v0.5` is. (I tagged it at the beginning of this week.)

assuming the review process is successful, wouldn't it make sense for me to cut a new release (`v0.6`) and update openjournals/joss-reviews#372 to point at it? (it currently refers to `v0.4`).
