### 2017-09-08 d4823f0
all, rootio: make Directory.Get return an error instead of a boolean
This CL modifies:
 Directory.Get(namecycle string) (Object, bool)

to:
 Directory.Get(namecycle string) (Object, error)

this allows to capture the context at the call site and thus reduces
code duplication when handling the error.
