### 2017-09-08 #372 @jpivarski
Nothing I checked was 0.4 specific, in fact "go get go-hep" got the latest master. (You might want to change that to the latest _release_, if possible.) I concur that the JOSS stamp should be on v0.5 or a new v0.6. There are release notes about fixes to the module I was testing in 0.5.

Also, I think I'm done with the review. It was a lot easier than I thought. I'm going to "close" this issue because I think it's done.

Congratulations! It's nice software! (And benchmark it on some uncompressed files!)
