### 2017-09-08 #372 @sbinet
> Congratulations! It's nice software! (And benchmark it on some uncompressed files!)
thanks! (and thanks for the tip: in my ACAT paper, I was actually quoting compressed and uncompressed perf numbers. I'll send a PR against [jblomer/iotools](https://github.com/jblomer/iotools) repo to get some "official" benchmark numbers.)

@jpivarski @arfon : is there anything else left to do on my side? apart from merging my `joss-paper` branch into `master` and tag `v0.6` ?
