### 2017-09-08 5c254ce
rootio/cmd/root-{dump,diff}: first stab
