+++ rootio/cmd/root-dump/main.go
@@ -1,0 +1,165 @@
+// Copyright 2017 The go-hep Authors.  All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+// root-dump dumps the content of a ROOT file, including the content of
+// the Trees (for all entries), if any.
+//
+// Example:
+//
+//  $> root-dump ./testdata/small-flat-tree.root
+//  >>> file[./testdata/small-flat-tree.root]
+//  key[000]: tree;1 "my tree title" (TTree)
+//  [000][Int32]: 0
+//  [000][Int64]: 0
+//  [000][UInt32]: 0
+//  [000][UInt64]: 0
+//  [000][Float32]: 0
+//  [000][Float64]: 0
+//  [000][Str]: evt-000
+//  [000][ArrayInt32]: [0 0 0 0 0 0 0 0 0 0]
+//  [000][ArrayInt64]: [0 0 0 0 0 0 0 0 0 0]
+//  [000][ArrayInt32]: [0 0 0 0 0 0 0 0 0 0]
+//  [000][ArrayInt64]: [0 0 0 0 0 0 0 0 0 0]
+//  [000][ArrayFloat32]: [0 0 0 0 0 0 0 0 0 0]
+//  [000][ArrayFloat64]: [0 0 0 0 0 0 0 0 0 0]
+//  [000][N]: 0
+//  [000][SliceInt32]: []
+//  [000][SliceInt64]: []
+//  [...]
+//
+//  $> root-dump -h
+//  Usage: root-dump [options] f0.root [f1.root [...]]
+//
+//  ex:
+//   $> root-dump ./testdata/small-flat-tree.root
+//   $> root-dump -deep=0 ./testdata/small-flat-tree.root
+//
+//  options:
+//    -deep
+//      	enable deep dumping of values (including Trees' entries) (default true)
+//
+package main // import "go-hep.org/x/hep/rootio/cmd/root-dump"
+
+import (
+	"flag"
+	"fmt"
+	"io"
+	"log"
+	"os"
+	"reflect"
+
+	"go-hep.org/x/hep/rootio"
+)
+
+var (
+	deepFlag = flag.Bool("deep", true, "enable deep dumping of values (including Trees' entries)")
+)
+
+func main() {
+	log.SetPrefix("root-dump: ")
+	log.SetFlags(0)
+
+	flag.Usage = func() {
+		fmt.Fprintf(os.Stderr, `Usage: root-dump [options] f0.root [f1.root [...]]
+
+ex:
+ $> root-dump ./testdata/small-flat-tree.root
+ $> root-dump -deep=0 ./testdata/small-flat-tree.root
+
+options:
+`,
+		)
+		flag.PrintDefaults()
+	}
+
+	flag.Parse()
+
+	if flag.NArg() == 0 {
+		flag.Usage()
+		log.Fatalf("need at least one input ROOT file")
+	}
+
+	for _, fname := range flag.Args() {
+		err := dump(os.Stdout, fname, *deepFlag)
+		if err != nil {
+			log.Fatalf("error dumping file %q: %v", fname, err)
+		}
+	}
+}
+
+func dump(w io.Writer, fname string, deep bool) error {
+	fmt.Fprintf(w, ">>> file[%s]\n", fname)
+	f, err := rootio.Open(fname)
+	if err != nil {
+		return err
+	}
+	defer f.Close()
+
+	for i, key := range f.Keys() {
+		obj, err := key.Object()
+		if err != nil {
+			return err
+		}
+		fmt.Fprintf(w, "key[%03d]: %s;%d %q (%s)\n", i, key.Name(), key.Cycle(), key.Title(), obj.Class())
+		if deep {
[…]
