### 2017-09-08 #372 @arfon
> Okay, so I'm not supposed to close it. Awaiting further instruction.

👍 thanks for being proactive by we don't close the issues until we've accepted the paper.
