### 2017-09-08 #372 @arfon
@sbinet - it looks like Pandoc can't find a couple of the references - could you take a look?

```
pandoc-citeproc: reference ROOT not found
pandoc-citeproc: reference Go not found
```
