### 2017-09-08 #372 @arfon
OK thanks. This is now looking good.

@sbinet - At this point could you merge the `joss-paper` branch into `master` and tag `v0.6`. Then please make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
