### 2017-09-08 #372 @sbinet
- `joss-paper` merged into `master`: `[DONE]`
- tag `v0.6`: `[DONE]`
- DOI: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.887895.svg)](https://doi.org/10.5281/zenodo.887895)

thanks! :)
